CSlideShowPic: split into GL/GLES/DX classes
