+++ xbmc/pictures/SlideShowPicture.h
@@ -12,13 +12,8 @@
 #include "threads/CriticalSection.h"
 #include "utils/ColorUtils.h"
 
-#include <string>
-#ifdef HAS_DX
-#include "guilib/GUIShaderDX.h"
-#include <wrl/client.h>
-#endif
-
 #include <memory>
+#include <string>
 
 class CTexture;
 
@@ -35,8 +30,10 @@ public:
     int length = 0;
   };
 
+  static std::unique_ptr<CSlideShowPic> CreateSlideShowPicture();
+
   CSlideShowPic();
-  ~CSlideShowPic();
+  virtual ~CSlideShowPic();
 
   void SetTexture(int iSlideNumber,
                   std::unique_ptr<CTexture> pTexture,
@@ -81,13 +78,17 @@ public:
   bool m_bIsComic;
   bool m_bCanMoveHorizontally;
   bool m_bCanMoveVertically;
+
+protected:
+  virtual void Render(float* x, float* y, CTexture* pTexture, UTILS::COLOR::Color color) = 0;
+
 private:
   void SetTexture_Internal(int iSlideNumber,
                            std::unique_ptr<CTexture> pTexture,
                            DISPLAY_EFFECT dispEffect = EFFECT_RANDOM,
                            TRANSITION_EFFECT transEffect = FADEIN_FADEOUT);
   void UpdateVertices(float cur_x[4], float cur_y[4], const float new_x[4], const float new_y[4], CDirtyRegionList &dirtyregions);
-  void Render(float* x, float* y, CTexture* pTexture, UTILS::COLOR::Color color);
+
   std::unique_ptr<CTexture> m_pImage;
 
   int m_iOriginalWidth;
@@ -132,8 +133,4 @@ private:
   bool m_bTransitionImmediately;
 
   CCriticalSection m_textureAccess;
-#ifdef HAS_DX
-  Microsoft::WRL::ComPtr<ID3D11Buffer> m_vb;
-  bool UpdateVertexBuffer(Vertex *vertices);
-#endif
 };
